iot_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
redis==5.0.1
celery==5.3.4
paho-mqtt==1.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
requests==2.31.0
python-dotenv==1.0.0

# IoT Protocol Support (Optional - Install as needed)
# CoAP - Constrained Application Protocol
aiocoap==0.4.7

# AMQP - Advanced Message Queuing Protocol
pika==1.3.2

# Bluetooth Low Energy
bleak==0.21.1

# Zigbee (Optional - requires additional dependencies)
# zigpy-znp==0.11.0

# LoRaWAN (Optional)
# lorawan==0.3.0

# Matter/Thread (Optional - in development)
# matter-server==1.5.0

# WebSocket support
# websockets==12.0
